Use BigDecimal instead of Double
[#177988195]

+++ src/test/java/org/codeforamerica/shiba/output/caf/NonHourlyJobIncomeInformationTest.java
@@ -4,6 +4,8 @@ import org.codeforamerica.shiba.Money;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
+import java.math.BigDecimal;
+
 import static org.assertj.core.api.Assertions.assertThat;
 
 class NonHourlyJobIncomeInformationTest {
@@ -15,7 +17,7 @@ class NonHourlyJobIncomeInformationTest {
             "EVERY_MONTH,1",
             "IT_VARIES,1"
     })
-    void shouldCalculateGrossMonthlyIncome(String payPeriod, Double income) {
+    void shouldCalculateGrossMonthlyIncome(String payPeriod, BigDecimal income) {
         assertThat(new NonHourlyJobIncomeInformation(payPeriod, "1", 0, null).grossMonthlyIncome())
                 .isEqualTo(new Money(income));
     }
